double space for version string
git-svn-id: svn+ssh://cdcvs.fnal.gov/cvs/projects/trace-svn/trunk@417 f0fafde3-ccc7-4eb4-bb98-6daa41875d3e

+++ include/trace.h
@@ -213,7 +213,7 @@ union trace_mode_u
 
 struct traceControl_s
 {
-    char           version_string[sizeof(int32_t)*16];
+    char           version_string[sizeof(int64_t)*16];
     uint32_t	   version;
     uint32_t       num_params;
     uint32_t       siz_msg;
